+++ src/format/pg-jsonl/index.js
@@ -19,6 +19,7 @@ export default {
 
   hierarchy: false,
   hyperEdges: false,
+  multiEdges: true,
 
   datatypes: "JSON scalars",