+++ src/format/pg/index.js
@@ -20,6 +20,7 @@ export default {
 
   hierarchy: false,
   hyperEdges: false,
+  multiEdges: true,
 
   datatypes: "JSON scalars",